# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "GabomaPay",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://oask.ca/api"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "CLIENT & ADMINISTRATEUR_AUTH"
                ],
                "summary": "Authentifie l'utilisateur avant de se connecter",
                "description": "",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "CLIENT & ADMINISTRATEUR_AUTH"
                ],
                "summary": "Deconnecte l'utilisateur",
                "description": "",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "schema": "Bearer",
                                        "type": "string",
                                        "example": "89er4186gjazuihhiZIOJreioiouEIOJIOZERF814879AE8FEA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "CLIENT & ADMINISTRATEUR_AUTH"
                ],
                "summary": "Enregister un utisaterur",
                "description": "",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/forgot": {
            "post": {
                "tags": [
                    "CLIENT & ADMINISTRATEUR_AUTH"
                ],
                "summary": "Mot de passe oublié",
                "description": "",
                "operationId": "03aca08208571c2a43bcdeebb40304ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "bgi@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/reset": {
            "post": {
                "tags": [
                    "CLIENT & ADMINISTRATEUR_AUTH"
                ],
                "summary": "Renitialiser le mot de passe",
                "description": "",
                "operationId": "8ae8881c540d1f799293343fde43c530",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "ufghzruqgioqzruogrjcuuUIBYRCUR?61154600"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}